도끼 패킷, 스타트 패킷 수정
컨트롤러 수정

+++ Server/Server/02.Network/TimerEvent.h
@@ -12,7 +12,7 @@ enum class TimerEvent
 
 struct Timer
 {
-	uint16_t playerId;
+	int32 playerId;
 	std::chrono::system_clock::time_point wakeUpTime;
 	TimerEvent _eventType;
